+++ resources/views/add-players-to-game.blade.php
@@ -16,7 +16,7 @@
         <div class="col-lg-8 mx-auto p-3 py-md-5">
 
             <main>
-                <form action="{{ route('game.add-players.process', ['game_id' => $game_id]) }}" method="POST" class="col-12 col-md-4 col-lg-4 mx-auto p-2">
+                <form action="{{ route('game.add-players.action', ['game_id' => $game_id]) }}" method="POST" class="col-12 col-md-4 col-lg-4 mx-auto p-2">
 
                     @csrf
